fixed bug: event queue was returning int instead of double for eventQ::firstTime

+++ MASH-CPP/inst/include/MASHcpp/HUMAN-EventQ.hpp
@@ -36,7 +36,7 @@ public:
   };
 
   // return time of first event
-  int firstTime(){
+  double firstTime(){
     return(EventQ[0]["tEvent"]);
   };
